+++ src/pages/api/contact-feedback/[id].ts
@@ -6,6 +6,7 @@ import ContactFeedbackService from '@/services/contactFeedback.service';
 import { ContactFeedback as IContactFeedback } from '@/interfaces/contactFeedback';
 import { Types } from 'mongoose';
 import { isBoolean } from '@/utils/validTypes';
+import HashHandlerService from '@/services/hash.service';
 
 type Response = {
   data: string | IContactFeedback;
@@ -23,9 +24,9 @@ export default async function handler(
     return;
   }
 
-  await dbConnect();
-
   try {
+    await dbConnect();
+
     const objectId = new Types.ObjectId(id as string);
     const foundFeedback = await ContactFeedback.findById(objectId);
 
@@ -37,8 +38,13 @@ export default async function handler(
     if (method === httpMethods.get) {
       res.status(200).json({ data: ContactFeedbackService.fromServer(foundFeedback) as IContactFeedback });
     } else if (method === httpMethods.patch) {
-      const { reviewed } = req.body;
+      const verify = await HashHandlerService.verifyAdminToken(req.headers.authorization);
+      if (!verify) {
+        res.status(401).json({ data: 'Invalid access token or role' });
+        return;
+      }
 
+      const { reviewed } = req.body;
       if (!isBoolean(reviewed)) {
         res.status(400).json({ data: 'Reviewed flag should be boolean' });
         return;
@@ -47,6 +53,12 @@ export default async function handler(
       const updated = await ContactFeedback.findByIdAndUpdate(objectId, { reviewed }, { new: true });
       res.status(201).json({ data: ContactFeedbackService.fromServer(updated) as IContactFeedback });
     } else if (method === httpMethods.delete) {
+      const verify = await HashHandlerService.verifyAdminToken(req.headers.authorization);
+      if (!verify) {
+        res.status(401).json({ data: 'Invalid access token or role' });
+        return;
+      }
+
       await ContactFeedback.findByIdAndDelete(objectId);
       res.status(200).json({ data: 'Contact feedback has been sucessfully deleted' });
     } else {
